+++ scout/commands/load/base.py
@@ -14,6 +14,7 @@ from .region import region as region_command
 from .user import user as user_command
 from .report import delivery_report as delivery_report_command
 from .report import cnv_report as cnv_report_command
+from .report import coverage_qc_report as coverage_qc_report_command
 from .exons import exons as exons_command
 
 LOG = logging.getLogger(__name__)
@@ -35,4 +36,5 @@ load.add_command(research_command)
 load.add_command(variants_command)
 load.add_command(delivery_report_command)
 load.add_command(cnv_report_command)
+load.add_command(coverage_qc_report_command)
 load.add_command(exons_command)