+++ scout/demo/__init__.py
@@ -49,3 +49,5 @@ delivery_report_path = pkg_resources.resource_filename("scout", delivery_report_
 
 cnv_report_file = "demo/cancer_cnv_report.pdf"
 cnv_report_path = pkg_resources.resource_filename("scout", cnv_report_file)
+coverage_qc_report_file = "demo/cancer_coverage_qc_report.html"
+coverage_qc_report_path = pkg_resources.resource_filename("scout", coverage_qc_report_file)