+++ CHANGELOG.md
@@ -8,6 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Added a page about migrating potentially breaking changes to the documentation
 - markdown_include in development requirements file
+- Coverage and quality report to cancer view
 ### Fixed
 - ACMG classification page crashing when trying to visualize a classification that was removed
 - Pretty print HGVS on gene variants (URL-decode VEP)